abstract base class usage
