+++ src/aligned_textgrid/mixins/tiermixins.py
@@ -27,42 +27,7 @@ class TierMixins:
         if hasattr(self, "sequence_list"):
             raise IndexError(f"{type(self).__name__} tier with name"\
                              f" {self.name} has empty sequence_list")
-        raise AttributeError(f"{type(self).__name__} is not indexable.")
-
-    def __contains__(self, item):
-        return item in self.sequence_list
-    
-    def __getitem__(self, idx):
-        return self.sequence_list[idx]
-    
-    def __iter__(self):
-        self._idx = 0
-        return self
-
-    def __len__(self):
-        return len(self.sequence_list)
-
-    def __next__(self):
-        if self._idx < len(self.sequence_list):
-            out = self.sequence_list[self._idx]
-            self._idx += 1
-            return(out)
-        raise StopIteration
-
-    def index(
-            self, 
-            entry
-        ) -> int:
-        """Return index of a tier entry
-
-        Args:
-            entry (SequencePoint |SequenceInterval):
-                A SequenceInterval or a PointInterval to get the index of.
-
-        Returns:
-            (int): The entry's index
-        """
-        return self.sequence_list.index(entry)
+        raise AttributeError(f"{type(self).__name__} is not indexable.")\
     
 
 class TierGroupMixins:
@@ -71,9 +36,6 @@ class TierGroupMixins:
     Attributes:
         []: Indexable and iterable
     """
-
-    def __contains__(self, item):
-        return item in self.tier_list
     
     def __getattr__(
             self,
@@ -91,35 +53,7 @@ class TierGroupMixins:
         
         if len(match_list) < 1:
             raise AttributeError(f"{type(self).__name__} has no attribute {name}")
-
-    def __getitem__(
-            self,
-            idx: int|list
-    ):
-        if type(idx) is int:
-            return self.tier_list[idx]
-        if len(idx) != len(self):
-            raise Exception("Attempt to index with incompatible list")
-        if type(idx) is list:
-            out_list = []
-            for x, tier in zip(idx, self.tier_list):
-                out_list.append(tier[x])
-            return(out_list)
-    
-    def __iter__(self):
-        self._idx = 0
-        return self
-
-    def __len__(self):
-        return len(self.tier_list)
-
-    def __next__(self):
-        if self._idx < len(self.tier_list):
-            out = self.tier_list[self._idx]
-            self._idx += 1
-            return(out)
-        raise StopIteration
-    
+        
     @property
     def name(self):
         if self._name:
